# server/javaWebProject/src/main/java/com/webshop/app/dto/AuthRequestDTO.java
package com.webshop.app.dto;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@AllArgsConstructor
@NoArgsConstructor
@Builder
public class AuthRequestDTO {

    private String email;
    private String password;
}