# swagger.yaml
openapi: 3.0.0
info:
  title: Webshop API
  description: |
    API za Webshop aplikaciju koja omogućuje pregled i kupnju proizvoda, upravljanje korisničkim računima,
    te administraciju proizvoda i kategorija.
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Lokalni razvojni server
paths:
  /api/login:
    post:
      tags:
        - Autentikacija
      summary: Prijava korisnika
      description: Autentikacija korisnika i izdavanje JWT tokena
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRequestDTO'
      responses:
        '200':
          description: Uspješna prijava
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserInfoDTO'
        '401':
          description: Neuspješna autentikacija
  /api/register:
    post:
      tags:
        - Autentikacija
      summary: Registracija korisnika
      description: Registracija novog korisnika u sustav
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterDTO'
      responses:
        '200':
          description: Uspješna registracija
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponseDTO'
        '400':
          description: Greška pri registraciji
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponseDTO'
  /api/refreshToken:
    post:
      tags:
        - Autentikacija
      summary: Obnova JWT tokena
      description: Korištenje refresh tokena za dobivanje novog JWT pristupnog tokena
      responses:
        '200':
          description: Uspješna obnova tokena
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserInfoDTO'
        '400':
          description: Nevažeći refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponseDTO'
...
  /api/categories:
    get:
      tags:
        - Kategorije
      summary: Dohvati sve kategorije
      description: Dohvaća sve kategorije proizvoda
      responses:
        '200':
          description: Uspješan dohvat
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CategoryDTO'
                  
    post:
      tags:
        - Kategorije
      summary: Kreiraj novu kategoriju
      description: Stvara novu kategoriju proizvoda
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryDTO'
      responses:
        '200':
          description: Kategorija uspješno kreirana
        '400':
          description: Greška pri kreiranju kategorije
        '401':
          description: Neautorizirani pristup
  /api/categories/{id}:
    get:
      tags:
        - Kategorije
      summary: Dohvati kategoriju po ID-u
      description: Dohvaća detalje o pojedinoj kategoriji
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: ID kategorije
      responses:
        '200':
          description: Uspješan dohvat
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryDTO'
        '404':
          description: Kategorija nije pronađena
          
    put:
      tags:
        - Kategorije
      summary: Ažuriraj kategoriju
      description: Ažurira postojeću kategoriju
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: ID kategorije
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryDTO'
      responses:
        '200':
          description: Kategorija uspješno ažurirana
        '400':
          description: Greška pri ažuriranju kategorije
        '401':
          description: Neautorizirani pristup
          
    delete:
      tags:
        - Kategorije
      summary: Izbriši kategoriju
      description: Briše kategoriju po ID-u
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: ID kategorije
      responses:
        '200':
          description: Kategorija uspješno obrisana
        '400':
          description: Greška pri brisanju kategorije
        '401':
          description: Neautorizirani pristup
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    AuthRequestDTO:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    RegisterDTO:
      type: object
      required:
        - firstName
        - lastName
        - phoneNumber
        - email
        - password
        - repeatPassword
      properties:
        firstName:
          type: string
          minLength: 1
          maxLength: 100
          description: Ime korisnika. Mora imati između 1 i 100 znakova.
        lastName:
          type: string
          minLength: 1
          maxLength: 100
          description: Prezime korisnika. Mora imati između 1 i 100 znakova.
        phoneNumber:
          type: string
          description: Broj mobitela korisnika.
        email:
          type: string
          format: email
          description: Email adresa korisnika.
        password:
          type: string
          minLength: 6
          description: Lozinka. Mora imati najmanje 6 znakova.
        repeatPassword:
          type: string
          minLength: 6
          description: Potvrda lozinke. Mora se podudarati s lozinkom.
    UserInfoDTO:
      type: object
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        role:
          $ref: '#/components/schemas/ApplicationUserRole'
    ApplicationUserRole:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
          description: Naziv uloge, npr. ROLE_USER, ROLE_ADMIN
    MessageResponseDTO:
      type: object
      properties:
        message:
          type: string
...
    CategoryDTO:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
...